Replaced dense storage of digits w/ VecDeque of digits
As I was developing this library & looking ahead to adding division,
it became clear that pushing to the beginning of the list was going
to be a more common operation than I'd originally anticipated, and
that I'd need to use a double-ended queue for digits.

While it was possible to extend the original approach (which was
to use bitwise operations for dense storage of digits within a usize)
to have better support for double-ended operations (eg by keeping
track of the index of both the first and last digit, instead of just
the last digit), this would incur a significant complexity cost.

I benchmarked a simple test emulating addition with carry using
both the dense storage approach and the far simpler approach of
using a VecDeque<u8> where each u8 contains the digit's value.
The VecDeque approach was ~3x faster, though of course consumed
considerably more memory (eg 2x the memory in base 10).

Given the huge savings in complexity and the considerably gains in
speed, I decided to opt for the VecDeque.