:Refactored to remove Base, replace with usize
I decided I need to prune complexity to get a working version, and I am skeptical
that the precomputation approach I was taking with Base would yield savings, anyway.

I removed Base and am now representing bases with simply a usize, and the utilities
Base depended on.

When the library is feature-complete & the optimization phase can begin in earnest,
I will benchmark using the precomputation tables, and contemplate adding Base back in.