Improved is_digit() to detect when the fractional part is 0
Improved is_digit() to detect when power is negative but all of the fractional
digits are 0, and updated the integer conversion functions to filter out any
negative powers they're given.

This required the deque_range feature, which is currently unstable. However
it will be stabilized in the 1.51.0, which will be released very soon.