Implemented *Assign traits, using hacks
Implemented:
- AddAssign
- SubAssign
- MulAssign
- ShrAssign
- (ShlAssign was implemented previously)

For:
- Number
- usize
- isize

I had to use some pretty gross hacks to get this to work. I needed it
to work for both Number and &mut Number, so that I could still use
the operations from within the Assign traits of other operations, and
to provide a sane API to the end user. However I didn't want to use
something like Borrow<Number> to accomplish this as it would impose
an unnecessary runtime cost.

The solution was to implement the operations for &mut Number, and then
implement the traits for owned Numbers by casting them to &mut Number.
The compiler should optimize out this indirection.

I tried to use macros to clean this up, but due to the nuances of
`self` in macros (https://danielkeep.github.io/tlborm/book/mbe-min-non-identifier-identifiers.html),
this ended up being harder than was really worth it. I will probably
still use a marco to clean this up a bit better in a refactor, but
for now, it is working & unblocks me to work on division.